extract client part from navbar, theme fix,absolute imports
